Updated checkout layout and improved design responsiveness
The commit introduces a new section in the checkout page for a custom checkout experience, resulting in modification of HTML and SCSS files. It also includes enhancement of responsiveness for different display sizes, delivering an optimized checkout layout for users. Changes in the report.html file also indicate an update in the date of Webpack Bundle Analyzer report.